Fixed bug for reseeding of GSL random number generator by the GssaStoich.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/dh_branch@3358 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: d3885e69d1f669c0e1954d509228975d760aa0b6

+++ ksolve/GssaStoich.cpp
@@ -403,8 +403,8 @@ void GssaStoich::reinit( const Eref& e, ProcPtr p )
 		unsigned int start = 0;
 		for ( unsigned int i = 0; i < Shell::numProcessThreads(); ++i )
 		{
-			unsigned long j = 0;
-			unsigned long seed = mtrand() * ( j - 1 );
+			unsigned int j = 0;
+			unsigned int seed = mtrand() * ( j - 1 );
 			gsl_rng_set( randNumGenerators_[i], seed );
 			vector< unsigned int >& mi = meshIndex_[ i ];
 			unsigned int end =
